Reorganize admin controllers and navbar items
