+++ src/Controller/Admin/TagController.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Controller;
+namespace App\Controller\Admin;
 
 use App\Entity\Tag;
 use App\Entity\TagTranslation;
