+++ templates/partials/_navbar.html.twig
@@ -1,6 +1,6 @@
 <nav class="navbar bg-dark fixed-top navbar-expand-lg" data-bs-theme="dark">
     <div class="container-fluid">
-        <span class="navbar-brand">Blog Demo</span>
+        <a class="navbar-brand" href="{{ path('app_home') }}">Blog Demo</a>
         <button class="navbar-toggler" type="button" data-bs-toggle="collapse" data-bs-target="#navbarSupportedContent"
                 aria-controls="navbarSupportedContent" aria-expanded="false" aria-label="Toggle navigation">
             <span class="navbar-toggler-icon"></span>
@@ -13,20 +13,28 @@
                     </a>
                 </li>
                 {% if is_granted('ROLE_ADMIN') %}
-                    <li class="nav-item">
-                        <a class="nav-link active" aria-current="page" href="{{ path('app_user_index') }}">
-                            {% trans %}Users{% endtrans %}
+                    <li class="nav-item dropdown">
+                        <a class="nav-link dropdown-toggle" href="#" role="button" data-bs-toggle="dropdown"
+                           aria-expanded="false">
+                            {% trans %}Administration{% endtrans %}
                         </a>
-                    </li>
-                    <li class="nav-item">
-                        <a class="nav-link active" aria-current="page" href="{{ path('app_tag_index') }}">
-                            {% trans %}Tags{% endtrans %}
-                        </a>
-                    </li>
-                    <li class="nav-item">
-                        <a class="nav-link active" aria-current="page" href="{{ path('app_post_index') }}">
-                            {% trans %}Posts{% endtrans %}
-                        </a>
+                        <ul class="dropdown-menu dropdown-menu-end" data-bs-theme="light">
+                            <li>
+                                <a class="dropdown-item" href="{{ path('app_user_index') }}">
+                                    {% trans %}Users{% endtrans %}
+                                </a>
+                            </li>
+                            <li>
+                                <a class="dropdown-item" href="{{ path('app_tag_index') }}">
+                                    {% trans %}Tags{% endtrans %}
+                                </a>
+                            </li>
+                            <li>
+                                <a class="dropdown-item" href="{{ path('app_post_index') }}">
+                                    {% trans %}Posts{% endtrans %}
+                                </a>
+                            </li>
+                        </ul>
                     </li>
                 {% endif %}
                 <li class="nav-item dropdown">
