+++ translations/messages+intl-icu.en.json
@@ -67,5 +67,6 @@
     "Language": "Language",
     "English": "English",
     "Croatian": "Croatian",
-    "Go back home": "Go back home"
+    "Go back home": "Go back home",
+    "Administration": "Administration"
 }
