+++ translations/messages+intl-icu.hr.json
@@ -67,5 +67,6 @@
     "Language": "Jezik",
     "English": "Engleski",
     "Croatian": "Hrvatski",
-    "Go back home": "Vrati se na početnu"
+    "Go back home": "Vrati se na početnu",
+    "Administration": "Administracija"
 }
